+++ lib/phx_live_storybook/stories/attr.ex
@@ -1,10 +1,10 @@
 defmodule PhxLiveStorybook.Stories.Attr do
   @moduledoc """
-  An attr is one of your component attributes. Its structure mimics the
-  upcoming LiveView 0.18.0 declarative assigns.
+  An attr is one of your component attributes. Its structure mimics the LiveView 0.18.0 declarative
+  assigns.
 
-  Attributes declaration will populate the Playground tab of your storybook,
-  for each of your components.
+  Attributes declaration will populate the Playground tab of your storybook, for each of your
+  components.
 
   Supported keys:
   - `id`: the attribute id (required). Should match your component assign.
@@ -17,14 +17,12 @@ defmodule PhxLiveStorybook.Stories.Attr do
     * `:float` - any float
     * `:map` - any map
     * `:list` - a List of any arbitrary types
-    * `:block` - a block (you can only declare one)
-    * `:slot` - a slot (you can declare more than one)
     * Any struct module
   - `required`: `true` if the attribute is mandatory.
   - `default`: attribute default value.
   - `examples` the list or range of examples suggested for the attribute
-  - `values` the list or range of all possible examples for the attribute. Unlike examples, this option
-  enforces validation of the default value against the given list.
+  - `values` the list or range of all possible examples for the attribute. Unlike examples, this
+     option enforces validation of the default value against the given list.
   - `doc`: a text documentation for this attribute.
 
   """